pyproject.toml
[project]
name = "py-minecraft-sync"
version = "0.1.0"
description = ""
authors = [
    {name = "Patrick Romão",email = "patrickromao2001@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "dropbox (>=12.0.2,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
